Implement chain sync pipelining to maximise websocket bandwith
With this, I've also increased the size of the minimum allocation area
so that the garbage collector would run more often. This also has a
significant impact since A LOT of time is actually spent garbage
collecting stuff out of memory.

Reaching 6500 block/s which is even faster than the current
cardano-wallet :rocket: