remove ChainSync trace and add TxSubmission tracers
The ChainSync traces are relatively garbage / noise and really not helpful even for debugging.
Plus, having an explicit trace still causes the underlying tracer to make comparison with the
severity to see whether or not something needs to be printed.

On the contrary, the tracer for the tx submission is quite useful to have and provide valuable
information.